requirements.txt
amqp==1.4.9
anyjson==0.3.3
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
enum-compat==0.0.2
eventlet==0.21.0
greenlet==0.4.13
idna==2.7
kombu==3.0.37
lxml==4.2.1
mock==2.0.0
nameko==2.9.0
path.py==11.0.1
pbr==4.0.4
pkg-resources==0.0.0
python-decouple==3.1
PyYAML==3.12
requests==2.19.1
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
wrapt==1.10.11